DRF_CRUD_reg_login_auth_complete/requirements.txt
astroid==2.2.5
beautifulsoup4==4.7.1
bleach==3.1.0
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-environ==0.4.5
django-filter==2.1.0
django-filters==0.2.1
django-rest-swagger==2.2.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.1.5
idna==2.8
isort==4.3.17
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.3.1
lxml==4.3.2
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
# pkg-resources==0.0.0
PyJWT==1.7.1
pylint==2.3.1
pytz==2019.1
requests==2.21.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.8
sqlparse==0.3.0
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.4.3
webencodings==0.5.1
wrapt==1.11.1